Fix web cli
