+++ stograde/webapp/web_cli.py
@@ -1,3 +1,4 @@
+import functools
 import os
 from PyInquirer import prompt, style_from_dict, Token
 from typing import List, TYPE_CHECKING
@@ -8,7 +9,8 @@ from ..process_assignment import import_supporting, remove_supporting
 from ..process_file import process_file
 from ..process_file.file_result import FileResult
 from ..process_file.process_file import get_file
-from ..student.process_student import prepare_student
+from ..student.process_student import prepare_student_repo
+from ..toolkit.process_repos import process_parallel_repos
 
 if TYPE_CHECKING:
     from ..specs.spec import Spec
@@ -17,10 +19,12 @@ if TYPE_CHECKING:
 def launch_cli(base_dir: str,
                clean: bool,
                date: str,
+               no_progress_bar: bool,
                skip_repo_update: bool,
                spec: 'Spec',
                stogit_url: str,
-               students: List[str]) -> bool:
+               students: List[str],
+               workers: int):
     """Start the web grading CLI"""
     usernames = [
         '{} NO SUBMISSION'.format(student)
@@ -29,22 +33,28 @@ def launch_cli(base_dir: str,
         for student in students
     ]
 
-    for student in students:
-        prepare_student(student,
-                        stogit_url=stogit_url,
-                        do_clean=clean,
-                        do_clone=not skip_repo_update,
-                        do_pull=not skip_repo_update,
-                        do_checkout=True,
-                        date=date)
+    print('Loading repos. Please wait...')
+
+    with chdir(os.path.join(base_dir, 'students')):
+        single_repo = functools.partial(
+            prepare_student_repo,
+            stogit_url=stogit_url,
+            do_clean=clean,
+            do_clone=not skip_repo_update,
+            do_pull=not skip_repo_update,
+            do_checkout=True,
+            date=date)
+
+    process_parallel_repos(students=students,
+                           no_progress_bar=no_progress_bar,
+                           workers=workers,
+                           operation=single_repo)
 
     while True:
         student = ask_student(usernames)
 
         if not student or student == 'QUIT':
-            return False
-        elif student == 'LOG and QUIT':
-            return True
+            return
         elif 'NO SUBMISSION' in student:
             continue
 
@@ -67,7 +77,7 @@ def ask_student(usernames: List[str]) -> str:
             'type': 'list',
             'name': 'student',
             'message': 'Choose student',
-            'choices': ['QUIT', 'LOG and QUIT', *usernames]
+            'choices': ['QUIT', *usernames]
         }
     ]
